+++ src/api/Rulings.ts
@@ -1,3 +1,4 @@
+import Cached from "../util/Cached";
 import MagicQuerier, { List } from "../util/MagicQuerier";
 
 export interface Ruling {
@@ -8,22 +9,27 @@ export interface Ruling {
 
 class Rulings extends MagicQuerier {
 
+	@Cached
 	public async byId (id: string) {
 		return (await this.query<List<Ruling>>(["cards", id, "rulings"])).data;
 	}
 
+	@Cached
 	public async bySet (setCode: string, collectorNumber: string | number) {
 		return (await this.query<List<Ruling>>(["cards", setCode, `${collectorNumber}`, "rulings"])).data;
 	}
 
+	@Cached
 	public async byMultiverseId (id: number) {
 		return (await this.query<List<Ruling>>(["cards/multiverse", id, "rulings"])).data;
 	}
 
+	@Cached
 	public async byMtgoId (id: number) {
 		return (await this.query<List<Ruling>>(["cards/mtgo", id, "rulings"])).data;
 	}
 
+	@Cached
 	public async byArenaId (id: number) {
 		return (await this.query<List<Ruling>>(["cards/arena", id, "rulings"])).data;
 	}